Renamed I2CContainer.py to i2c_container.py and updated sources
