+++ plugins/lpdpower/lpdpower/quad.py
@@ -1,5 +1,5 @@
 from i2c_device import I2CException
-from I2CContainer import I2CContainer
+from i2c_container import I2CContainer
 from tca9548 import TCA9548
 from mcp23008 import MCP23008
 from ad7998 import AD7998
